1、添加主键标识 2、按照领域驱动设计的思想，将Data Transfer Object结构转变为Value Object结构。

+++ customer-management/src/main/java/com/szsc/customermanagement/entity/Company.java
@@ -13,7 +13,7 @@ import com.szsc.customermanagement.enums.LocationEnum;
 @ToString
 public class Company {
     @ExcelIgnore // 忽略Excel导入导出
-    private Integer companyId; // 主键字段
+    private Long companyId; // 主键字段
 
     @ExcelProperty("公司代码")
     private String companyCode;
